refactor(editor): Relocate useTelemetry and useStorage below the stores tier
`@n8n/composables` depends on `@n8n/stores`, so any store moving into
`@n8n/stores` that imports a composable needs `stores → composables` — a
build-fatal cycle, since `build` is topological (`dependsOn: ["^build"]`).

`useTelemetry` and `useStorage` are cycle-free: neither references
`@n8n/stores`. They were simply mis-placed. Both move to
`@n8n/frontend-utils`, which sits below the stores tier, and the old paths
(`@n8n/composables/useTelemetry`, `@/app/composables/useStorage`) stay as
re-export shims so no call site changes.

`useTelemetry` holds a module-level registered instance, so the relocated
copy plus a re-export is only safe while both paths resolve to one module
instance. Two tests now assert that: a build-level check that
`@n8n/composables` externalizes rather than inlines the dependency, and a
behavioural check in `editor-ui` that an instance registered through one
path is observed through the other.
