+++ packages/frontend/@n8n/frontend-utils/README.md
@@ -1,8 +1,8 @@
 # @n8n/frontend-utils
 
-A collection of framework-light utility functions shared across n8n's front-end
-packages — HTML sanitization, DOM helpers, and other pure utilities that are not
-Vue composables.
+A collection of framework-light helpers shared across n8n's front-end packages —
+HTML sanitization, DOM helpers, and the small composables that carry no
+component, store, or app coupling.
 
 ## Table of Contents
 
@@ -12,7 +12,12 @@ Vue composables.
 
 ## Features
 
-- **Pure utilities**: Stateless helpers with no component or store coupling.
+- **No upward coupling**: nothing here imports a component, a store, or the app.
+  This is the package's real boundary — it is what lets every other front-end
+  package, `@n8n/stores` included, depend on it.
+- **Leaf position**: depends only on framework libraries and shared type
+  packages (`@n8n/api-types`, `n8n-workflow`) — never on another
+  `packages/frontend` package, so importing it can never close a cycle.
 - **Consistency**: A single home for cross-package front-end utilities.
 - **Extensible**: A foundation for utilities as the front end grows.
