+++ packages/frontend/@n8n/frontend-utils/src/useStorage.test.ts
@@ -1,4 +1,5 @@
 import { nextTick } from 'vue';
+
 import { useStorage } from './useStorage';
 
 describe('useStorage', () => {
