+++ packages/frontend/@n8n/stores/package.json
@@ -39,6 +39,7 @@
     "@n8n/api-types": "workspace:*",
     "@n8n/constants": "workspace:*",
     "@n8n/frontend-constants": "workspace:*",
+    "@n8n/frontend-utils": "workspace:*",
     "@n8n/i18n": "workspace:*",
     "@n8n/permissions": "workspace:*",
     "@n8n/rest-api-client": "workspace:*",
